requirements.txt
asgiref==3.3.1
astroid==2.4.2
colorama==0.4.3
coverage==5.3
dj-database-url==0.5.0
Django==3.1.3
django-cors-headers==3.6.0
django-heroku==0.3.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
green==3.2.2
gunicorn==20.0.4
isort==5.6.4
lazy-object-proxy==1.4.3
lxml==4.5.2
mccabe==0.6.1
psycopg2==2.8.6
PyJWT==1.7.1
pylint==2.6.0
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
Unidecode==1.1.1
whitenoise==5.2.0
wrapt==1.12.1
